// src/app/shered/models/partida-ordinaria.models.ts
export interface partidaOrdinaria{
  idmovimiento: any;
  iddepartamento: any;
  nombreDepartamento: any;
  cuenta: any;
  unidades: any;
  descripcion: any;
  justificacion: any;
  precio: any;
  total:any;
  mes01: any;
  mes02: any;
  mes03: any;
  mes04: any;
  mes05: any;
  mes06: any;
  prorrateo: any;
  iva: any;
  ingreso: any;
  idivade: any;
  autorizada: any;
  semestre: any;
  julio: any;
  agosto: any;
  septiembre: any;
  octubre: any;
  noviembre: any;
  diciembre: any;
  esPartidaParcial: any;
  estatus: any;
  montoSolicitado: any;
  saldoFavor: any;
  tienePartidaCapturada: any;
}